chore(iast): fix iast span metrics
