+++ ddtrace/appsec/_iast/_taint_utils.py
@@ -7,6 +7,7 @@ from typing import Union
 
 from ddtrace.appsec._iast._taint_tracking._taint_objects import is_pyobject_tainted
 from ddtrace.appsec._iast._taint_tracking._taint_objects import taint_pyobject
+from ddtrace.appsec._iast.constants import DBAPI_INTEGRATIONS
 from ddtrace.internal.logger import get_logger
 from ddtrace.settings.asm import config as asm_config
 
@@ -513,6 +514,17 @@ class LazyTaintDict:
 
     def urlencode(self, safe=None):
         return self._taint(self._obj.urlencode(safe=safe), self._origin_value)
+
+
+def supported_dbapi_integration(integration_name):
+    return integration_name in DBAPI_INTEGRATIONS or integration_name.startswith(DBAPI_PREFIXES)
+
+
+def check_tainted_dbapi_args(args, kwargs, integration_name, method):
+    if supported_dbapi_integration(integration_name) and method.__name__ == "execute":
+        return len(args) and args[0] and is_pyobject_tainted(args[0])
+
+    return False
 
 
 if asm_config._iast_lazy_taint:
