+++ tests/appsec/iast/test_taint_utils.py
@@ -198,19 +198,18 @@ def test_checked_tainted_args(iast_context_defaults):
 
     # Returns False: Untainted first argument
     assert not check_tainted_dbapi_args(
-        args=(untainted_arg,), kwargs=None, tracer=None, integration_name="sqlite", method=cursor.execute
+        args=(untainted_arg,), kwargs=None, integration_name="sqlite", method=cursor.execute
     )
 
     # Returns False: Untainted first argument
     assert not check_tainted_dbapi_args(
-        args=(untainted_arg, tainted_arg), kwargs=None, tracer=None, integration_name="sqlite", method=cursor.execute
+        args=(untainted_arg, tainted_arg), kwargs=None, integration_name="sqlite", method=cursor.execute
     )
 
     # Returns False: Integration name not in list
     assert not check_tainted_dbapi_args(
         args=(tainted_arg,),
         kwargs=None,
-        tracer=None,
         integration_name="nosqlite",
         method=cursor.execute,
     )
@@ -219,24 +218,23 @@ def test_checked_tainted_args(iast_context_defaults):
     assert not check_tainted_dbapi_args(
         args=(tainted_arg,),
         kwargs=None,
-        tracer=None,
         integration_name="sqlite",
         method=cursor.executemany,
     )
 
     # Returns True:
     assert check_tainted_dbapi_args(
-        args=(tainted_arg, untainted_arg), kwargs=None, tracer=None, integration_name="sqlite", method=cursor.execute
+        args=(tainted_arg, untainted_arg), kwargs=None, integration_name="sqlite", method=cursor.execute
     )
 
     # Returns True:
     assert check_tainted_dbapi_args(
-        args=(tainted_arg, untainted_arg), kwargs=None, tracer=None, integration_name="mysql", method=cursor.execute
+        args=(tainted_arg, untainted_arg), kwargs=None, integration_name="mysql", method=cursor.execute
     )
 
     # Returns True:
     assert check_tainted_dbapi_args(
-        args=(tainted_arg, untainted_arg), kwargs=None, tracer=None, integration_name="psycopg", method=cursor.execute
+        args=(tainted_arg, untainted_arg), kwargs=None, integration_name="psycopg", method=cursor.execute
     )
